trying to follow conventional file organization methods, restructered
