+++ src/engine/mod.rs
@@ -1,10 +1,8 @@
 use neobridge_rust::RGB;
 use screenshots::image::{ImageBuffer, Pixel, Rgba};
 
-mod channel;
-
 use crate::{
-    color, term::CalculationOption
+    calc, term::CalculationOption
 };
 
 pub struct JellyRenderer {
@@ -15,7 +13,7 @@ pub struct JellyRenderer {
 
     depth: usize,
 
-    color_option: color::ColorOption,
+    color_option: calc::ColorOption,
     calc_option: CalculationOption,
 
     result: Vec<RGB>,
@@ -27,7 +25,7 @@ impl JellyRenderer {
         height: u32,
         n_of_leds: usize,
         depth: usize,
-        color_option: color::ColorOption,
+        color_option: calc::ColorOption,
         calc_option: CalculationOption,
     ) -> JellyRenderer {
         JellyRenderer {
@@ -49,7 +47,7 @@ impl JellyRenderer {
         let mut x: u32 = 0;
 
         // we want to get a certain amount of pixels at the bottom.
-        let mut channels: channel::ChannelStorage = channel::ChannelStorage::new(self.depth);
+        let mut channels: calc::ChannelStorage = calc::ChannelStorage::new(self.depth);
         for _row in 0..self.n_of_leds {
             // we are getting a column of RGB values for each LED on a strip.
             // for example, if I have a depth of 10,
@@ -74,8 +72,8 @@ impl JellyRenderer {
             );
 
             if !(self.calc_option.disable_color_operations) {
-                rgb = color::ColorOperation::set_saturation(&rgb, self.color_option.saturation);
-                rgb = color::ColorOperation::set_brightness(&rgb, self.color_option.brightness);
+                rgb = calc::ColorOperation::set_saturation(&rgb, self.color_option.saturation);
+                rgb = calc::ColorOperation::set_brightness(&rgb, self.color_option.brightness);
             }
 
             self.result.push(rgb);
